Restore the CAB mode on orientation change
Currently, the CAB mode is restored only when the Conversation List
Fragment (CLF) is available.  The CLF was needed earlier because we
grabbed views from the selected conversations. We don't anymore, so
the CLF dependency can go.

However, we cannot always restore the CAB mode, we might be in
conversation view mode on the phone. On orientation change, we want to
show subject view in the ActionBar, not CAB mode. So we need to check
the view mode before doing this.

Bug: 7742352 Selected message action bar icons (delete,start) are
             missing from Gmail inbox list after orientation change

Change-Id: If6dcae413f07fe708c3cd6c970b2245b58c0a23a