Improve 'up' tap highlighting
Leave our custom SnippetTextView as-is, but dynamically switch
from CUSTOM to TITLE mode when it is safe to do so. This fixes
the 'up' tap highlight while leaving our snippet measurement
logic intact and functional.

I've lined up the layout of SnippetTextView to be consistent
with the current framework TITLE mode, but there are no
guarantees these will stay consistent on future platforms.
Caveat emptor.

Use reflection on the new ActionBar 2-line flag as a way to
determine if 2-line TITLE mode is supported in the platform.

Bug: 8030707
Change-Id: If293c20a296b36265d6dbcf49095879093ddfbc7