Fix crashes related to sending or forwarding attachments.
Attachment contentUri will always be a content:// uri, not the file path.
SERVER_ATTACHMENT will generally be under Gmail.AUTHORITY while LOCAL_FILE will
have a non-Gmail uri.

Attachments as part of a draft that becomes updated into a message are now
viewable. They will be a LOCAL_FILE that points to the non-Gmail contentUri of
the uploaded file. They will not attempt to be downloaded again. They will have
a fake partId and uri so that they can be viewed.

We make sure thumbnailUri and contentUri are always populated with the correct
default contentUri (for SERVER_ATTACHMENT).

The cachedFileUri is now also used for SERVER_ATTACHMENT. It points of the uri
of the downloaded file.

Bug: 8076816
Bug: 7960494
Change-Id: Ia73144886c6a730987bf6fe7b2326e6820f4e2f6