Don't call setSnippetMode unnecessarily
When looking at a conversation from a search result, the
search action bar would enter "snippet" mode on every time that
the options menu is populated.

The problem is that once the layout has happened at least once,
there won't be another layout.  So the subsequent call to
setSnippetMode was causing the actionbar to switch back to custom
mode, where the up behavior isn't implemented

This will be fixed better in UR8, where onprepare options menu will
not be used to switch the actionbar modes

Bug: 8124697

cherry-pick of https://googleplex-android-review.googlesource.com/#/c/269773/

Change-Id: If5acb4129af1055477a706762535b39a637159c9