Start showing "from" in message details under header
When looking at a message in a conversation, the user is unable to quickly
copy the sender's email - especially when it is a long email 
address/unknown contact. 

A possible way to fix this is to always display the "From" subject/email 
address. This will allow for users to long press on the text in order to 
copy as much as is required (From both the name/email).

Changes viaDomain boolean logic in order to always display "from" between
the "date" and "reply-to" rows in the message detail.


Bug:8142748 Look for ways to copy sender email address in conversation 
view by long pressing on message details to copy text

Change-Id: I0da55258186d6b73d59d22cd07c96a7f95887b73