Suppress automatic update to mark read/viewed when marking unread
Every conversation shown in the CVF is marked read, because it has
been read by the user. When we mark messages unread, there is a race
between the marking unread, and this automatic marking read. In rare
cases, the cache can get the operations in reverse order:

1. User clicks on message
2. User marks unread (in a hurry)
3. Message is shown
4. CVF.onConversationSeen() is called, which
5.... marks message read and viewed.

The easiest fix is to suppress marking the message read when
markUnread() is called.

Bug: 7326925 Mark Unread Non-Functional When Viewing Message
Change-Id: If57f674f39b5dbf115a468fcd17656f9715394ed