Create Message Dialog upon long pressing the MessageHeader
When you're attempting to check and copy a sender's email address,
you should not have to remember or type out that person's entire
address. One way to approach this problem is to add in a dialog
that allows for an easy way to copy the email address. Upon some
user action, the dialog can then pop up for the user to copy the
sender's email address.

In this case, the user action will be a long press on the message
upper header in a Conversation View, such that a dialog with the
correct sender's email shows up (Unique to each message in the
conversation). The long press, however, will not affect conversations
that are collapsed i.e. the message header was tapped and the
message is not visible.

Bug:8142748

Change-Id: I0c736d89e6b1183ea46d852fd8574125931081ed