Support "Move to" functionality alongside "Change labels"
User cannot move an email from one defined label to another.
Change label is a hassle to use to only select one and a
menu item for moving is preferable.

This change introduces the "Move to" as part of the overflow
menu. It allows for users to move emails from folders, excluding
Starred/Important.

Bug: 8157593

Change-Id: Ib5ae9f04b52f3f5d0f2d2ddb590219f7733cf171