Created category for user-assignable folders (i.e. not starred/important)
UIProvider was missing a category that only allows for
certain labels to be viewed and assigned to by users
through some folder UI (Change label/Move To).

This change adds said constant into FolderCapabilities
related to when the user should be able to move a
conversation. Written as CAN_BE_USER_ASSIGNED.

Bug: 8157593

Change-Id: Id1143d70cde539528f065035e1a6e96797e4ae02