Preserve adapter on view mode changes
Currently, we stopListening() to conversation cursor changes on the
*start* of a view mode change from conversation mode to conv_list
mode.  This has the side effect of changing the CPA.getCount() from
<whatever> to 1, since we don't have a controller, and thus cannot get
the count.

Change stopListening() to do nothing, while preserving all callers.
This should fix the crash with the side effect of jank in rare cases
where a conversation cursor update comes in during the transition, and
knocks off all three conversations that the pager created fragments
for.

A subsequent cleanup cl will remove callers and beautify associated
code.

Bug: 8190963 Tap/Select back key before a message is read in Gmail
             conversation view shows IllegalStateException

Change-Id: I87252a09ea90dd43e107dc26a37b328ea01bb0ab