Improve null cursor checking
Currently, isPagingDisabled() calls getCursor().  In many methods, a
second call to getCursor() is made. The cursor in the two cases could
be different depending on the timing of the events as the controller
could have received an updated cursor, or the controller could have
vanished in the meanwhile.

Changing isPagingDisabled() to accept a cursor instead. As a side
benefit, we can eliminate (cursor == null) checks which is either:
A. redundant: since isPagingDisabled checks these
OR
B. incorrect: if the cursor was non-null in isPagingDisabled but null later.

Change-Id: I082b0907cefcce7f358fce87ce351b22a04a71cc