Enabled swipe to archive/delete to change immediately
There was an issue with swipe to archive/delete. The
steps are recorded with the bug. Essentially, the
original behavior is swipe to archive. Once the user
sets that to "no effect" through settings, he/she
can no longer swipe. However, once the user turned it
back on, swipe to archive still did not work.

Ths change adds "enableSwipe" to the "setSwipeAction"
method as, previously, only the action was changed,
but the actual swipe action was not being re-enabled
immediately.

Bug: 8203494

Change-Id: I3c64034b796b21cb1baa53440d26d59e6cde079a