Recognize low disk, and propagate the appropriate error code
If the device gets in low disk space mode, Gmail sync will potentially
never succeed till the device recovers from this mode. In such a case,
propagate the appropriate code up to the UI so it can accurately
message the user.

This is Unified Email's half of the code, where we actually set the
storage_low/storage_ok state.

Bug: 7662097 Low Disk causes "No connection" dialog in gmail
Change-Id: I08004aad821b6c4f17ed6179e04726df9ec1b8b4