Remove back button when Waiting For Sync
When waiting for sync (No Conversation List/Item shown),
there used to be a back arrow without any purpose attached
to the Gmail icon.

The change adds on consideration for the view mode being
"WAITING_FOR_ACCOUNT_INITIALIZATION" such that the action
bar icon does not present the back button.

Bug: 8142710
Change-Id: I9294b25d3bd13ab54c1bc6e5a5d54d0770bb97c8