Pop out of a draft after discarding it. (Part 2)
If the Activity housing the ConversationViewFragment is killed while
the ComposeActivity is up, interesting things happen when the
Activity is restored.

It restores its old Fragment, calls popOut() on it, but it is not
added, so it does not pop out.  Then the ConversationPagerAdapter
creates a new Fragment, but popOut() is never called on it.

Now we will tell new Fragments that they are in detached mode if they
are created after the adapter has already entered detached mode, to
give them an opportunity to pop out.

This change is not necessary in UR9, because of changes in
CursorStatus.isWaitingForResults(), and will be reverted there in
another CL.

Follow-up to Ieb27536edc2a716b76d005fcbcc250aa03693dfd

Bug: 8185448
Change-Id: If86b184f7b95ef656903cece4ff74bba2897b65b