optimize scroll perf
Switch date stamp back to a simple drawText(), as we had
switched to use a TextView to use font face spans, but this was
much slower. Since we are no longer using a custom font for
date, this is safe.

Tweak call ordering of measure/layout/setText of subject and
senders to be more like it was in UR7. Calling setText last
defers text layout until draw time, which is normally bad, but
when our list is already so far behind, it turns out there are
fewer draw calls than layout passes, and this saves some time.

Switch EllipsizedMultilineTextView from using its internal
layout to using a StaticLayout to determine the chop point. I
found this slightly faster for my data. I think this was because
most of my subject+snippets require StaticLayout vs.
BoringLayout, and even checking isn't worth it.

Bug: 8236797
Change-Id: I8684ed7cd58d9dbbaff1df4721a823a59635171c