Switched alert dialog to context menu for long press on message header
Originally, long clicking on the upper message header led
to the creation of an alert dialog that would display
the copy email address option. The issues with that were
not limited to improper truncation and poor UI choice.

The change borrows from WebViewContextMenu to create a
context menu on long pressing on the message header when
that message is expanded. The options present for the new
popup are copy and send mail to address.

Note, I will be fixing the chips side as well.

Bug: 8254505

Change-Id: Ia2eb0759e14f8b800324f96f7e0df27cff1d6f77