fix spurious 'mark unread' upon archive/delete on tablet
On tablet, when auto-advance is set to list, destructive actions
were having the side effect of marking the following
conversation read because the pager adapter no longer stops
listening during the transition. Let's bring back that behavior,
since it's generally a good thing to avoid changes during the
2-pane mode transition, but tweak it to placate ViewPager's new
dataset change requirement. When we stop listening, rather than
de-facto and unnecessarily switching to singleton mode, keep the
cursor around and use it as normal.

On the off-chance that we actually miss a real dataset change
during the transition, save off and use a frozen version of the
dataset count to placate ViewPager if it happens to ask for the
count during the transition.

Bug: 8272000
Change-Id: I9f52a50903eb5a7eb5c665b2573eb9b2d0d91c67