pop out of conversation view when dealing with singleton drafts
Previous CLs (Ieb2753, If86b18) popped out of conversation mode
when the visible conversation goes missing from the cursor AND
the conversation's message count goes to zero.

Sadly, Gmail reassigns IDs to outbound unsynced singleton draft
conversations upon sync, so the sync makes the current draft
seem like it's being deleted, which triggers detached mode.

A longer-term fix is to prevent those IDs from ever changing,
but for now, change the scope of the original fix. Rather than
pop out when message count goes to zero, do it when a
conversation that starts out as a singleton draft goes missing.

Now that the message cursor count is not used to trigger the
pop-out, simplify the code to cut the fragment out of the loop.

Bug: 8185448
Change-Id: I0bf78b11b93c71b13ad3ffa560a4ee0e75a9e542