Service to add arbitrary interaction logging to Gmail.
MailLogService provides logging to circular buffers for a given tag.
The application adds logging to these circular buffers. When a
bugreport is captured, the last fifty lines of logs per tag is printed
out on the bugreport.

A compile-time boolean can disable this entire functionality. Even if
it is enabled, the user needs to increase the log level using adb
shell setprop to start the service.  If the service is not running,
logging is still carried out, though it will never be dumped.

This is for development only. It needs to be turned off before
release.

Bug: 8293525 Add debugging information for Gmail

Change-Id: Id59a9ff9a88a40e1afdfd31f0441b9531e797d4d