start munging wide <div> and <table> elements
Modify HTML content that is too wide if the experimental flag
is on.

Find <table> elements with a width wider than the body and
remove the widths in the hope that the table would otherwise
fit. Content authors often use this trick to prevent their
content from growing too wide in a very wide desktop browser
window.

Failing that, find all <div> elements where a width is
specified that is wider than the body, and turn the width into
a max-width.

Bug: 7400516
Change-Id: Ie4d3c2c4c50a191c30ee590dce71bc9bf6de8268