Toast timeout for non-swipe archive
Archive/delete/undo toasts stay forever at the moment
until a user action is done that switches fragments
or some other view-changing event.

The change adds in a delayed runnable to hide the toast
after 15 seconds (Can be tested and changed as necessary).

***NOTE: This is NOT a change to the swipe to delete/archive,
which the bug is directly dealing with. However, since the
features are the same, I included it for tracking purposes.
I'll be working on the archive timeout change in the
meantime.

Bug: 8176151
Change-Id: Idcb4aaf0eb62c207c659156d8c116ce5141fb76e