(Experimental) Make Drawer list sectioned and show recents
Drawer previously was a flattened view and mail spinner
was the only dialog with recent labels. This change has
disabled recent labels in the spinner and changed the
drawer folder list to show sections (system/recent/all).

Note, functionality has not been removed in order to
consider Email functionality - I've only disabled the
spinner as is and changed the drawer to a sectioned one
(SECTIONS_AND_RECENT_FOLDERS_ENABLED in FolderListFrag).

Bug: 8373282
Change-Id: If53712478402abeb37715844c05305a90bf60cc5