Load ref message for edit draft.
When user replies to a message, save it as draft, open it again, and
toggle the "reply" vs. "reply all" drop down, we were not populating
the "to" and "cc" fields with the sender and recipients of the original
email because mRefMessage was not being populated and left as null.

The fix is to use an additional loader to query for the reference message
that the draft was replying to.  Because the refMessageId is stored
in database as the server message id, and there's no existing URI for
querying message by server message id, added such a URI to UIProvider.

Bug 8295718

Change-Id: If6815bb272131a1ab1209ac9a595de9eb43aa2f4