Add sync status bar
User triggered manual refresh will trigger this new sync status bar
to appear on the top of the conversation list.  It will show a progress
spinner indicating a sync is in progress until sync is finished.
If sync was successful but no messages were found, we also report that
in the status bar.

Bug 8351953

Change-Id: I71d77ee3f2f4606082792f68bf5993281a60fcef