Fix NPE in WidgetService
Because we changed the format of widget intents to use uris (folderUri
and folderConversationListUri), old widgets (installed in UR8) and
any intents that are persisted (by the framework) were triggering this
NPE since they don't have the uri extras set.  This is a temporary work
around where we still use the old folder extra in the intent.  A proper
fix would be to migrate the old intents to the new format.

Bug 8479183

Change-Id: If7eeb8e6ca84db302c32becabdee9a0ed05b0916