fix zero-size conversation lists from a view conversation intent
Execute the conversation list transaction immediately.
Small scope of a fix for now... I'll expand later if it's safe.

This causes the add-list operation to insert a view immediately, so
ViewPager + FragmentStatePagerAdapter don't get a chance to pick up this add-
list operation and execute it during measure(). That was causing the
beloved layout-in-layout problem of zero-sized view sub-trees.

Bug: 8483995
Change-Id: Idf2e1ce734d2593113b86b1c7a779757eb8d1358