Double checkbox tap zone
Checkbox needed more space as it was getting squeezed
to death on the left + drawer interfered with the left
tap zone.

See if there's other variations of lengths that help
you better/a little less. Doubling the touch zone
effectively allows the click pretty far out to the
right. Of course, once people play around with this
and give feedback, it'll help out.

Bug: 8195514
Change-Id: Ib0822fa9ad0a216b9f57d24a714d57bd5830b3e2