fix blank conversation view on tablet when launching from intent
If a 2-pane layout pass finished before the FIRST_FOLDER loader returned
results, the layout pass triggered an early showConversation() operation
before mConversationToShow is set by the FIRST_FOLDER loader.

The later showConversation() in the FIRST_FOLDER onLoadFinished did not
re-run the operation, because it assumed a transition would run that
showConversation had to wait for. The transition does not actually run
because TwoPaneLayout debounces consecutive requests to change the view
mode.

So...
* don't switch modes on FIRST_FOLDER load unnecessarily
* FIRST_FOLDER load was triggering a mode change to LIST, so move that
  into loadAccountInbox
* don't have 2-pane ask ViewMode to see if the mode changed, have it
  ask TPL instead

Add some helpful optional logging, and a facility to ObjectCursorLoader
to slow down race conditions for debugging purposes.

Bug: 8408885
Change-Id: If4672c0f6c4426f810de33f65694119d678bacfb