Allow back from top-level
Back had the implicit assumption that the parent had a folder list as
well. With a drawer, the top level is no longer a fragment.  Check if
the parent of the current folder list was also a hierarchical list. If
not, go back to the default inbox for the account.

Bug: 8474211 Folder List Fragment for Hierarchical folders back broken

Change-Id: I0c79f496e2a327ce7529e11e26e1f0dd096730e9