fix shorter-than-expected leave-behind view heights
Make leave-behind heights variable, dependent on the
ConversationItemView that originated it.
Make the background on a SwipeableCIV not a separate view, but a simple
background. Also simplify the conditions under which it occurs-- it just
appears whenever translateX is non-zero. This covers all swiping and
animating cases handily, without the need for extra listeners.

Remove an old workaround that gave ListView on 1-pane an opaque background-

- it was there to mask FolderListFragments that stuck around, but those
are relegated to the drawer now. This reduces list overdraw by one.

Bug: 8570880
Change-Id: I2e2ea875d4ab58b5048d41025b8a2c48a3d63536