Ignore updates to the current account in the drawer
When we get updates to the current account (with updated settings, for instance)
we don't want to restart any loaders.  The loader should only be restarted
when the current account changes from null -> non-null or when it changes
from accountA -> accountB.  In all other cases, we should change our
internal representation but do nothing else.

Oversight on my part, I should have noticed this before.

Bug: 8558083 Gmail: Flickering after selecting any option from sliding
             panel

Change-Id: Ia5a256bae6deddbc6819b407161cf35720a87cd8