Switch out of wait mode correctly
The burden of transitioning out of wait fragment is on the
implementation of hideWaitForInitialization(). It is correctly handled
on phone, but not on tablet.

On tablet, if we have not transitioned out to the correct mode, we
should set the mode to conversation list mode, which is the only valid
mode to set at this point.

Bug: 8529007 Compose, Search, Manage labels and Settings are missing
             from menu bar in list view on Nakasi after adding an
             account

Bug: 8427487 Blank page is displayed upon opening Gmail 4.3 (UR8) on
             ICS tablets

Change-Id: I779969187e1aaf0b84697354544b6f0440e93578