Disappear, white label color (UnifiedEmail)
There was an inconsistency in what the default colors
used across the app were. Labels have the color #f1f5ec to
begin with and when remove color is called on the web
interface, the default background color is set to #eeeeee.

Changed the colors value for default foreground for sake
of consistency.

Bug: 8612176
Change-Id: I72173fa20dd0478fc8318188c75e338c88d4b819