misc contact photo optimizations
* Only invalidate the contact image region on async load.
* Calculate hash codes by hand to save autoboxing of ints.
* Don't parse Folder color strings at draw time.
* Switch ConcurrentHashMap to SynchronizedMap. No need for concurrent
  write support, as we only have one writer thread. Write perf is much
  better.

No-op stale photo requests more correctly now. Previously, a canvas that
was re-used and happened to have the same person in it would apply the
photo even if the dimensions didn't match. Now, PhotoManager makes a
record of the canvas's "generation" at request time, so the results can
later be rejected entirely if the canvas generation differs.

Bug: 8051779
Change-Id: I31c3fd0c3bcc10641f264b6e30eb133cc0bf77b3