Run handle/resetNotificationActions on UI thread.
Fixes b/8703644 by making the bodies of
handleNotificationActions and resetNotificationActions
post onto the UI thread for all of their work and thus
fixes the ConcurrentModificationException of
mNotificationTempDeleted. Also fixes b/8608274
because cacheValue is no longer being called
on a non-UI thread in these two functions.

Also contains a partial fix for notification
actions being undone while the conversation
view pager was open. Now we only update the
ConversationCursor's internal state when the
undo notification is tapped rather than whenever
it is dismissed.

Change-Id: I937dd7f7a0f7284a6f66b982fe363a63c3618f3f