Only show unread count in widgets when non-zero.
Previously, the text view that held the unread count
in the widget was always visible. Now, we show it whenever
the unread count is not 0 and set the visibility to GONE
whenever the unread count is 0. This gives extra space
to the folder name and account name whenever the unread
count is 0 (much of the time). Fixes b/8743911.

Change-Id: I6b6c4ff4d21f1883f0bef953a8ff3022b8ec1e3a