tablet visual tweaks. more caret.
bump the star up a bit on 'normal' and 'spacious' layouts.
add back the caret for tablets.
because the caret is drawn as part of a list item, move the edge effect
line into the list item as well so it draws underneath the caret.
get the divider between selected list items appearing again by using a
9-patch instead of a solid color.

Fix an issue on portrait tablet where the last opened conversation would
stay blue in the list. The activated state wasn't being cleared on view
mode change.

Change-Id: I616f5d2b9eb9506995a74a4772ab7f50f95b2a33