Update settings per latest UX spec.
Fixes b/8773002. Turns Remove Action into Archive & delete
actions with three options. Setting that option changes
the title of the text in the swipe to setting below it.

This CL adds strings, changes around the preferences to support
the new tri-state setting, and updates the action bar based on
the setting (to show archive, delete, or archive & delete and
push move to into the overflow). An accompanying CL does the
rest of the work.

Change-Id: I17d8ada9bc9f16b5e7b5a8abd8c55e8708773991