Reflect settings changes immediately in actionbar.
Previously, we could return to conversation view and
still see the old settings for show archive/delete/archive&delete.
Now, in AbstractActivityController.onResume, we invalidate the options
menu so that it is always up to date when we return to the conversation
view or any other view mode, for that matter. Partially fixes b/8856931.

Change-Id: Ief1c909d91f710b211b5ea897fcc1dc122418d43