Make sender photos match between conv list & view.
Fixes b/8770455 for a device. Other devices may
return slightly different images but on a single device,
the sender images in the conversation list will match the
sender images in the conversation view.

The code in SenderInfoLoader was refactored into
one big static method called loadContactPhotos. That
method has a switch for whether to decode the bitmaps
or just return the byte array. Then, in PhotoManager,
some of the use of Object was changed to use their actual
types (Long or String). Finally, loadPhotos in ContactPhotoManager
was wired up to use the static method in SenderInfoLoader.
As a result, the images now match.

Change-Id: Ia8490ddfa11d5e3233d6ef6a642dabf2a8213af2