Prevent infinite loop in ContactPhotoManager.
Fixes b/8878905. PhotoManager prunes requests
after they have been operated on (whether or not
they have been been satisfied or we're displaying
a stock photo). A bug on the refactor made it so
we never pruned requests that were not satisfied.
This fix ensures all requests that have been operated
on get pruned.

Change-Id: Ida1dc90ab0f05b5b7ab90633cbb9749cae3c1577