Properly handle loading conversation list.
Fixes b/8846152. We now properly check whether
the drawer is animating when getting the conversation
list back from the loader. We now check the actual
state of the DrawerLayout (by storing the last state
from the DrawerListener) and querying it when we get
into onLoadFinished. If the drawer is not in STATE_IDLE,
we set a flag to query when the drawer state becomes STATE_IDLE.

Change-Id: I1ec35858f2659f01103d5b0505ebbda92eb49442