Fix underlying change processing during reset
If underlyingChanged is called in resetCursor(), the Controller will
call refresh(), and that does nothing if a refresh task is present. It
turns out that the refresh task is expected to be present but inert in
resetCursor(), so ensure that it is nulled out BEFORE resetCursor().
This also better matches the calling order pattern in setCursor().

Fixing this fixes some cases where mRefreshRequired was stuck at true,
preventing new/more mail from appearing.

Bug: 8961122
Change-Id: I6696cfe5d7d1a612379ff2a3c11069f754807ab2