Adding drag-state drawable to folder item icons.
Fixes b/8986628. Previously we had two states for
icons in the folder list (activated and default).
We now have a third state for when the app is in
drag mode and the folder is disabled which uses
a lighter grey icon. To make this work, we use
a custom drawable state (state_drag_mode) for
when drag is entered. Another CL will use this
new drag state for changing folder text color as well.

Change-Id: I280b187a454064f1c50f4ef62db5c8ba86c1f4d6
(cherry picked from commit 47fb86e95836e1a47bd784f6866c8c798b44ab6b)