not all respond() calls require locking
The UI thread uses respond() to signal user-visibility on each list
item, so we definitely don't want to block the UI thread on those kinds
of respond() calls. Add an options bitfield whose inaugural value is a
bit that hints that the cursor position must be safely moved prior to
respond().

This fixes some big hiccups that result when syncing + scrolling with
the respond CLs yesterday.

Bug: 9032162
Change-Id: I1c927453b9f474c89861a7295de1e318279e3a49