fix refresh ProgressBar halo
Turns out 'wrap_content' means little/nothing to ProgressBar: it ends up
sized as 48dp per the minHeight/maxHeight in the framework styles. This
causes the drawable to draw to a 48dp height, so funny artifacts aren't
surprising.

Force a fixed height for the refresh progress bar based on the custom
asset height.

Bug: 9101567
Change-Id: Id7949895bc757b3522d059e4fba4c1713c2a1da8