Fix hierarchical FolderListFragments
We were never using most of the args we set.

There are multiple active FLFs (drawer and subfolder). Both of these
are notified of folder changes, and then they check if that folder
exists in their cursor.  If not, they switch you back to the account's
default inbox.  However, the cursor for the drawer only has top level
folders, so this breaks.  Now, we have a full folder list cursor in
each FLF, and this is used for that check.

Bug: 8575001
Change-Id: I47f17fb9c5d108b2388db5ba918f6413a018b414