fix chained activity leak on compose rotation
Each rotation was adding a link to the reference chain via:
ComposeActivity
to mSavedInstanceState
to (view hierarchy state)
to TextView.SavedState
to SpannedString
to EasyEditSpanController (a span)
to Editor
to InputContentType
to OnEditorActionListener
which is either a RecipientEditTextView, or ComposeActivity, both of
which point to the old ComposeActivity.

This fix breaks the Activity->Activity chain by severing
mSavedInstanceState when it's no longer useful, but there is still a
1-deep leak due to TextView saving spans with references to the old
Activity.

Bug: 9189358
Change-Id: Ia4f6ba60fc0a0a3f4f3fd04f81832808d38d3f3c