don't keep a reference to view hierarchy saved state
It doesn't seem to fix the 1-deep Activity leak in ComposeActivity, but
at least it eliminates one reason why the previous ComposeActivity
sticks around after you rotate.

Since we do the unusual method of keeping the saved instance state
around as a member, we can at least restrict this state to just keeping
OUR state around so the view hierarchy state can be lost. Wrap all
activity state in an inner Bundle to do this.

Bug: 9212605
Change-Id: Id08e5347981724030af11108ad7937f79daf00b4