fix jank during auto-advance animation
A mode change due to auto-advance may trigger an animation. To ensure
that the animation is totally smooth, defer the pending destructive
action until after the mode change transition is complete.

When the transition doesn't run, we execute mAutoAdvanceOp immediately
by ensuring that mAutoAdvanceOp is set before showConversation(). We
assume that the no-op transition signals completion within the
showConversation() method call.

Bug: 9077521
Change-Id: I3f9f509ebd3fb0d9dc54ff7532ac7e843b6e9baf