Export Eml viewer.
There's really no reason why it shouldn't be
and having it exported prevents b/9334785 from occurring.
Additionally, so that we only use the app's eml viewer,
we check if the attachment is an eml file, and if so,
just open the viewer directly rather than letting the
framework disambiguate (which causes an activity chooser
to show when we have multiple activities that can accept
eml files).

Change-Id: I98e13f894c9218d62049b16e0f4c20c77501fcc5